pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
Flask = "^2.2.0"
google-api-python-client = "^2.95.0"
google-auth-oauthlib = "^1.0.0"
xds-protos = "^0.0.11"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}
urllib3 = "1.26.15"
poetry = {url="https://storage.googleapis.com/poetry-bundles/poetry-1.1.15-py2.py3-none-any.whl"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"